requirements.txt
streamlit==1.34.0
gradio_client==0.14.0
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.30.0 